refactor: streamline ReadyEndpointsCache and remove unused cache code
Rename `NewInformerBackedEndpointsCache` to `NewReadyEndpointsCacheWithInformer`
and move it to `ready_endpoints_cache.go` as it now returns a
`ReadyEndpointsCache`.

Since the scaler's queue pinger no longer listens for endpoint changes,
the `EndpointsCache` interface and `FakeEndpointsCache` have been removed
to simplify the codebase.
