+++ CHANGELOG.md
@@ -43,6 +43,7 @@ This changelog keeps track of work items that have been completed and are ready 
 
 - **General**: Fix `roundToNDigits` using `math.Round` instead of `math.Floor` to correctly round small negative floating-point errors to zero ([#1483](https://github.com/kedacore/http-add-on/pull/1483))
 - **Interceptor**: Remove configurable `KEDA_HTTP_ENDPOINTS_CACHE_POLLING_INTERVAL_MS` and hardcode EndpointSlice informer resync to 60m ([#1485](https://github.com/kedacore/http-add-on/pull/1485))
+- **Scaler**: Remove unused `KEDA_HTTP_SCALER_DEPLOYMENT_INFORMER_RSYNC_PERIOD` env var ([#1533](https://github.com/kedacore/http-add-on/pull/1533))
 
 ### Deprecations
