--- pkg/k8s/endpoints_cache_informer_test.go
@@ -1,45 +1,0 @@
-package k8s
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	discov1 "k8s.io/api/discovery/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/tools/cache"
-)
-
-func TestEndpointSliceFromDeleteObj_DirectObject(t *testing.T) {
-	r := require.New(t)
-	slice := &discov1.EndpointSlice{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "svc-slice",
-			Namespace: "testns",
-		},
-	}
-
-	got, err := endpointSliceFromDeleteObj(slice)
-	r.NoError(err)
-	r.Equal(slice, got)
-}
-
-func TestEndpointSliceFromDeleteObj_TombstoneValue(t *testing.T) {
-	r := require.New(t)
-	slice := &discov1.EndpointSlice{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "svc-slice",
-			Namespace: "testns",
-		},
-	}
-
-	got, err := endpointSliceFromDeleteObj(cache.DeletedFinalStateUnknown{Obj: slice})
-	r.NoError(err)
-	r.Equal(slice, got)
-}
-
-func TestEndpointSliceFromDeleteObj_InvalidTombstonePayload(t *testing.T) {
-	r := require.New(t)
-
-	_, err := endpointSliceFromDeleteObj(cache.DeletedFinalStateUnknown{Obj: "not-an-endpointslice"})
-	r.Error(err)
-}
