+++ pkg/k8s/ready_endpoints_cache_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/stretchr/testify/require"
 	discov1 "k8s.io/api/discovery/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/tools/cache"
 	"k8s.io/utils/ptr"
 )
 
@@ -226,6 +227,43 @@ func TestHasAnyReadyEndpoint_NilReadyIsReady(t *testing.T) {
 		},
 	}
 	r.True(hasAnyReadyEndpoint(slice), "nil Ready should be treated as ready per K8s API spec")
+}
+
+// --- endpointSliceFromDeleteObj tests ---
+
+func TestEndpointSliceFromDeleteObj_DirectObject(t *testing.T) {
+	r := require.New(t)
+	slice := &discov1.EndpointSlice{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "svc-slice",
+			Namespace: "testns",
+		},
+	}
+
+	got, err := endpointSliceFromDeleteObj(slice)
+	r.NoError(err)
+	r.Equal(slice, got)
+}
+
+func TestEndpointSliceFromDeleteObj_TombstoneValue(t *testing.T) {
+	r := require.New(t)
+	slice := &discov1.EndpointSlice{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "svc-slice",
+			Namespace: "testns",
+		},
+	}
+
+	got, err := endpointSliceFromDeleteObj(cache.DeletedFinalStateUnknown{Obj: slice})
+	r.NoError(err)
+	r.Equal(slice, got)
+}
+
+func TestEndpointSliceFromDeleteObj_InvalidTombstonePayload(t *testing.T) {
+	r := require.New(t)
+
+	_, err := endpointSliceFromDeleteObj(cache.DeletedFinalStateUnknown{Obj: "not-an-endpointslice"})
+	r.Error(err)
 }
 
 // --- helpers ---
